+++ History.md
@@ -1,7 +1,11 @@
 # History
 
+## 1.0.4 / 2017-06-26
+- signup
+- passwd -> password
+
 ## 1.0.2 / 2017-06-26
-- findUser updateUser updateUserExt updatePasswd signon encryptPasswd checkPasswd
+- findUser updateUser updateUserExt updatePassword signon encryptPassword checkPassword
 
 ## 1.0.0 / 2017-06-24
 - create